Refactorig MongoDb
